Error checks

+++ src/dbmstodspi/query_execution/scheduling/plan_evaluator_interface.hpp
@@ -55,6 +55,6 @@ class PlanEvaluatorInterface {
       const std::map<char, int>& cost_of_columns, double streaming_speed,
       double configuration_speed)
       -> std::tuple<std::vector<std::vector<ScheduledModule>>,
-                    std::vector<ScheduledModule>, int, int> = 0;
+                    std::vector<ScheduledModule>, long, long> = 0;
 };
 }  // namespace orkhestrafs::dbmstodspi
